+++ src/queries/index.js
@@ -2,3 +2,4 @@ export * from './useMarcSource';
 export * from './useAuthoritySourceFiles';
 export * from './useAuthorityLinkingRules';
 export * from './useAuthorityLinksCount';
+export * from './useLinkSuggestions';